// web/src/api/generated/models/PreferenceSentiment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const PreferenceSentiment = {
    LOVE: 'LOVE',
    LIKE: 'LIKE',
    NEUTRAL: 'NEUTRAL',
    DISLIKE: 'DISLIKE',
    AVOID: 'AVOID'
} as const;
export type PreferenceSentiment = typeof PreferenceSentiment[keyof typeof PreferenceSentiment];


export function instanceOfPreferenceSentiment(value: any): boolean {
    for (const key in PreferenceSentiment) {
        if (Object.prototype.hasOwnProperty.call(PreferenceSentiment, key)) {
            if (PreferenceSentiment[key as keyof typeof PreferenceSentiment] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PreferenceSentimentFromJSON(json: any): PreferenceSentiment {
    return PreferenceSentimentFromJSONTyped(json, false);
}

export function PreferenceSentimentFromJSONTyped(json: any, ignoreDiscriminator: boolean): PreferenceSentiment {
    return json as PreferenceSentiment;
}

export function PreferenceSentimentToJSON(value?: PreferenceSentiment | null): any {
    return value as any;
}

export function PreferenceSentimentToJSONTyped(value: any, ignoreDiscriminator: boolean): PreferenceSentiment {
    return value as PreferenceSentiment;
}